♻️ Divide into partial modules (utils.js)
